fix(review): move _skill_ok/_skill_fail from conftest to tests/server/_helpers.py
