+++ tests/server/_helpers.py
@@ -1,0 +1,34 @@
+"""Shared test builder utilities for tests/server/."""
+
+from __future__ import annotations
+
+from autoskillit.core import SkillResult
+from autoskillit.core.types import RetryReason
+
+
+def _skill_ok(report_text: str = "## Bug Report\ndetails") -> SkillResult:
+    return SkillResult(
+        success=True,
+        result=report_text,
+        session_id="sid",
+        subtype="success",
+        is_error=False,
+        exit_code=0,
+        needs_retry=False,
+        retry_reason=RetryReason.NONE,
+        stderr="",
+    )
+
+
+def _skill_fail() -> SkillResult:
+    return SkillResult(
+        success=False,
+        result="",
+        session_id="",
+        subtype="error",
+        is_error=True,
+        exit_code=1,
+        needs_retry=False,
+        retry_reason=RetryReason.NONE,
+        stderr="something went wrong",
+    )
